# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GCRF Library Management System API
  version: 1.0.0
  description: |
    RESTful API documentation for GCRF (国创睿峰) Intelligent Library Management System.

    This system provides comprehensive library management capabilities including:
    - Authentication and authorization
    - Book catalog management
    - Reader management
    - Circulation (borrow, return, renew)
    - System administration

    **Architecture**: Microservices with Spring Cloud Gateway

    **Base URL**: All requests go through the API Gateway at `http://localhost:8080`

    **Authentication**: JWT Bearer token required for most endpoints (see Authentication section)

  contact:
    name: GCRF Library Team
    email: support@gcrf-library.com
  license:
    name: Proprietary
    url: https://gcrf-library.com/license
servers:
  - url: http://localhost:8080
    description: Development server (Local)
  - url: https://api.gcrf-library.com
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/books:
    get:
      tags: [Books]
      summary: List books with pagination
      description: Query book catalog with search, filters and pagination
      parameters:
        - name: pageNum
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          description: Page size
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: keyword
          in: query
          description: Search keyword (title, author, ISBN)
          schema:
            type: string
        - name: classificationCode
          in: query
          description: Classification code filter
          schema:
            type: string
        - name: publisher
          in: query
          description: Publisher filter
          schema:
            type: string
        - name: status
          in: query
          description: Book status filter
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
      responses:
        '200':
          description: Book list retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedBookResponse'
        '400':
          $ref: '#/components/responses/ValidationError'

    post:
      tags: [Books]
      summary: Create new book
      description: Add a new book to the catalog (Admin/Librarian only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreateRequest'
      responses:
        '200':
          description: Book created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BookDetailVO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/books/{id}:
    get:
      tags: [Books]
      summary: Get book details
      description: Retrieve detailed information of a specific book
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Book ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Book details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BookDetailVO'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [Books]
      summary: Update book
      description: Update book information (Admin/Librarian only)
      parameters:
        - name: id
          in: path
          required: true
          description: Book ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookUpdateRequest'
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BookDetailVO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    delete:
      tags: [Books]
      summary: Delete book
      description: Soft delete a book from catalog (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          description: Book ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Book deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ========================================
  # Reader Management APIs
  # ========================================
...
  /api/v1/circulation/borrow:
    post:
      tags: [Circulation]
      summary: Borrow a book
      description: Create a borrowing record and decrease book inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowRequest'
      responses:
        '200':
          description: Book borrowed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CirculationRecord'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/circulation/return/{recordId}:
    post:
      tags: [Circulation]
      summary: Return a book
      description: Complete a borrowing record and increase book inventory
      parameters:
        - name: recordId
          in: path
          required: true
          description: Circulation record ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Book returned successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CirculationRecord'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/circulation/renew/{recordId}:
    post:
      tags: [Circulation]
      summary: Renew a book
      description: Extend the borrowing period
      parameters:
        - name: recordId
          in: path
          required: true
          description: Circulation record ID
          schema:
            type: integer
            format: int64
        - name: renewDays
          in: query
          description: Number of days to extend (default 30)
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 90
      responses:
        '200':
          description: Book renewed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CirculationRecord'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/circulation/reader/{readerId}:
    get:
      tags: [Circulation]
      summary: Get reader's circulation records
      description: Query circulation history for a specific reader
      parameters:
        - name: readerId
          in: path
          required: true
          description: Reader ID
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          description: Record status filter (0=Borrowed, 1=Returned, 2=Overdue)
          schema:
            type: integer
            enum: [0, 1, 2]
      responses:
        '200':
          description: Circulation records retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/CirculationRecord'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ========================================
  # Role Management APIs
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication. Include the token in the Authorization header:
        ```
        Authorization: Bearer <your-jwt-token>
        ```
...
  schemas:
...
    BookCreateRequest:
      type: object
      required:
        - isbn
        - title
        - author
        - publisher
        - totalQuantity
      properties:
        isbn:
          type: string
          pattern: '^(97[89])?\d{9}[\dX]$'
          description: ISBN-10 or ISBN-13
          example: "9787115551234"
        title:
          type: string
          maxLength: 500
          example: "Spring Boot实战"
        subtitle:
          type: string
          maxLength: 500
          example: "从入门到精通"
        author:
          type: string
          maxLength: 200
          example: "张三"
        translator:
          type: string
          maxLength: 200
          example: "李四"
        publisher:
          type: string
          maxLength: 200
          example: "人民邮电出版社"
        publishDate:
          type: string
          format: date
          example: "2023-06-01"
        edition:
          type: string
          maxLength: 50
          example: "第1版"
        pages:
          type: integer
          minimum: 1
          example: 520
        price:
          type: number
          format: decimal
          minimum: 0
          example: 89.00
        binding:
          type: string
          maxLength: 50
          example: "平装"
        language:
          type: string
          maxLength: 50
          default: "Chinese"
          example: "Chinese"
        classificationCode:
          type: string
          maxLength: 50
          example: "TP312"
        subjectKeywords:
          type: string
          maxLength: 500
          example: "Java, Spring Boot, 微服务"
        description:
          type: string
          maxLength: 2000
          example: "本书全面介绍Spring Boot框架的核心概念和实战技巧..."
        coverUrl:
          type: string
          format: uri
          maxLength: 500
          example: "https://cdn.gcrf-library.com/covers/book123.jpg"
        totalQuantity:
          type: integer
          minimum: 1
          example: 10
        availableQuantity:
          type: integer
          minimum: 0
          example: 8
    BookUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
        author:
          type: string
          maxLength: 200
        publisher:
          type: string
          maxLength: 200
        publishDate:
          type: string
          format: date
        price:
          type: number
          format: decimal
          minimum: 0
        description:
          type: string
          maxLength: 2000
        coverUrl:
          type: string
          format: uri
        totalQuantity:
          type: integer
          minimum: 0
        availableQuantity:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
    BookDetailVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        isbn:
          type: string
          example: "9787115551234"
        title:
          type: string
          example: "Spring Boot实战"
        subtitle:
          type: string
          example: "从入门到精通"
        author:
          type: string
          example: "张三"
        translator:
          type: string
          example: "李四"
        publisher:
          type: string
          example: "人民邮电出版社"
        publishDate:
          type: string
          format: date
          example: "2023-06-01"
        edition:
          type: string
          example: "第1版"
        pages:
          type: integer
          example: 520
        price:
          type: number
          format: decimal
          example: 89.00
        binding:
          type: string
          example: "平装"
        language:
          type: string
          example: "Chinese"
        classificationCode:
          type: string
          example: "TP312"
        subjectKeywords:
          type: string
          example: "Java, Spring Boot, 微服务"
        description:
          type: string
          example: "本书全面介绍Spring Boot框架的核心概念和实战技巧..."
        coverUrl:
          type: string
          format: uri
          example: "https://cdn.gcrf-library.com/covers/book123.jpg"
        totalQuantity:
          type: integer
          example: 10
        availableQuantity:
          type: integer
          example: 8
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
          example: "ACTIVE"
        createdAt:
          type: string
          format: date-time
          example: "2024-10-12T08:30:00"
        updatedAt:
          type: string
          format: date-time
          example: "2024-10-28T15:20:00"
...
    BorrowRequest:
      type: object
      required:
        - readerId
        - bookId
      properties:
        readerId:
          type: integer
          format: int64
          description: Reader ID
          example: 1
        bookId:
          type: integer
          format: int64
          description: Book ID
          example: 10
    CirculationRecord:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        readerId:
          type: integer
          format: int64
          example: 1
        bookId:
          type: integer
          format: int64
          example: 10
        borrowDate:
          type: string
          format: date-time
          example: "2024-10-28T14:30:00"
        dueDate:
          type: string
          format: date-time
          example: "2024-11-27T14:30:00"
        returnDate:
          type: string
          format: date-time
          nullable: true
          example: null
        renewCount:
          type: integer
          example: 0
        status:
          type: integer
          description: "0=Borrowed, 1=Returned, 2=Overdue"
          example: 0
        overdueDays:
          type: integer
          example: 0
        fine:
          type: integer
          description: Fine amount in cents (分)
          example: 0
...
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          description: Response code (200 for success)
          example: 200
        message:
          type: string
          description: Response message
          example: "操作成功"
        data:
          description: Response data (varies by endpoint)
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds
          example: 1698765432000
        traceId:
          type: string
          description: Request trace ID for debugging
          example: "trace-abc123xyz"
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: Error code
          example: 400
        message:
          type: string
          description: Error message
          example: "参数验证失败"
        data:
          nullable: true
        timestamp:
          type: integer
          format: int64
          example: 1698765432000
        traceId:
          type: string
          example: "trace-abc123xyz"
    PageResult:
      type: object
      properties:
        records:
          type: array
          items: {}
          description: List of records
        total:
          type: integer
          format: int64
          description: Total number of records
          example: 100
        pageNum:
          type: integer
          description: Current page number (1-based)
          example: 1
        pageSize:
          type: integer
          description: Page size
          example: 10
        pages:
          type: integer
          description: Total number of pages
          example: 10
    PagedBookResponse:
      allOf:
        - $ref: '#/components/schemas/PageResult'
        - type: object
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/BookDetailVO'
...
  responses:
    UnauthorizedError:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing token
              value:
                code: 401
                message: "未提供认证令牌"
                data: null
                timestamp: 1698765432000
            invalidToken:
              summary: Invalid token
              value:
                code: 401
                message: "令牌无效或已过期"
                data: null
                timestamp: 1698765432000
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                code: 403
                message: "权限不足，无法访问该资源"
                data: null
                timestamp: 1698765432000
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                code: 404
                message: "请求的资源不存在"
                data: null
                timestamp: 1698765432000
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                code: 400
                message: "参数验证失败：用户名不能为空"
                data: null
                timestamp: 1698765432000
...